requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
Faker
drf-yasg
django-filter
djangorestframework-simplejwt
django-cors-headers
gunicorn
python-dotenv
openpyxl
pandas